# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "BookingIn": {
        "properties": {
          "coach_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coach Id"
          },
          "duration_minutes": {
            "maximum": 1440,
            "minimum": 30,
            "title": "Duration Minutes",
            "type": "integer"
          },
          "resource_id": {
            "format": "uuid",
            "title": "Resource Id",
            "type": "string"
          },
          "session_type": {
            "$ref": "#/components/schemas/SessionType"
          },
          "starts_at": {
            "format": "date-time",
            "title": "Starts At",
            "type": "string"
          }
        },
        "required": [
          "resource_id",
          "starts_at",
          "duration_minutes",
          "session_type"
        ],
        "title": "BookingIn",
        "type": "object"
      },
...
      "ErrorOut": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
...
      "QuoteOut": {
        "properties": {
          "provisional": {
            "description": "cel puțin un tarif este DE_STABILIT",
            "title": "Provisional",
            "type": "boolean"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/SegmentOut"
            },
            "title": "Segments",
            "type": "array"
          },
          "total": {
            "description": "bani (RON × 100)",
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "provisional",
          "segments"
        ],
        "title": "QuoteOut",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "SessionUserAuth": {
        "in": "cookie",
        "name": "sessionid",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "API-ul central Jungle Padel. Erorile: `{\"error\": {\"code\", \"params\"}}`.",
    "title": "Jungle Padel API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/bookings": {
      "post": {
        "operationId": "jungle_bookings_api_create_booking",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingOut"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "SessionUserAuth": []
          }
        ],
        "summary": "Create Booking",
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/bookings/availability": {
      "get": {
        "description": "Occupied intervals per resource for one club-time day. No personal data (R-012).",
        "operationId": "jungle_bookings_api_availability",
        "parameters": [
          {
            "in": "query",
            "name": "location",
            "required": true,
            "schema": {
              "title": "Location",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "day",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Day",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailabilityOut"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Availability",
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/bookings/mine": {
      "get": {
        "operationId": "jungle_bookings_api_my_bookings",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BookingOut"
                  },
                  "title": "Response",
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "SessionUserAuth": []
          }
        ],
        "summary": "My Bookings",
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/bookings/quote": {
      "get": {
        "description": "The price before booking (R-052), with the per-30-minute breakdown.",
        "operationId": "jungle_bookings_api_get_quote",
        "parameters": [
          {
            "in": "query",
            "name": "resource_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Resource Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "starts_at",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "Starts At",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "duration_minutes",
            "required": true,
            "schema": {
              "title": "Duration Minutes",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "session_type",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "description": "R-044: shown on the court screen; changeable until check-in (Q29).",
                  "enum": [
                    "official_match",
                    "training",
                    "lesson",
                    "tournament",
                    "challenge",
                    "free_rental",
                    "event"
                  ],
                  "title": "SessionType",
                  "type": "string"
                }
              ],
              "default": "free_rental"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteOut"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Get Quote",
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/bookings/waitlist": {
      "get": {
        "operationId": "jungle_bookings_api_my_waitlist",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WaitlistOut"
                  },
                  "title": "Response",
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "SessionUserAuth": []
          }
        ],
        "summary": "My Waitlist",
        "tags": [
          "bookings"
        ]
      },
      "post": {
        "operationId": "jungle_bookings_api_join_waitlist",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaitlistOut"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "SessionUserAuth": []
          }
        ],
        "summary": "Join Waitlist",
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/bookings/waitlist/{entry_id}/leave": {
      "post": {
        "operationId": "jungle_bookings_api_leave_waitlist",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Entry Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkOut"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "security": [
          {
            "SessionUserAuth": []
          }
        ],
        "summary": "Leave Waitlist",
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/bookings/{booking_id}/cancel": {
      "post": {
        "operationId": "jungle_bookings_api_cancel_booking",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingOut"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "security": [
          {
            "SessionUserAuth": []
          }
        ],
        "summary": "Cancel Booking",
        "tags": [
          "bookings"
        ]
      }
    },
...
    "/api/v1/classes": {
      "get": {
        "operationId": "jungle_bookings_api_list_classes",
        "parameters": [
          {
            "in": "query",
            "name": "location",
            "required": true,
            "schema": {
              "title": "Location",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ClassOut"
                  },
                  "title": "Response",
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "List Classes",
        "tags": [
          "classes"
        ]
      }
    },
...
    "/api/v1/classes/mine": {
      "get": {
        "operationId": "jungle_bookings_api_my_enrollments",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EnrollmentOut"
                  },
                  "title": "Response",
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "SessionUserAuth": []
          }
        ],
        "summary": "My Enrollments",
        "tags": [
          "classes"
        ]
      }
    },
...
  },
  "servers": []
}